+++ packages/core/src/Cardano/util/estimateStakePoolAPY.ts
@@ -1,4 +1,5 @@
-import { Percent, StakePoolEpochRewards } from '../types';
+import { Percent } from '@cardano-sdk/util';
+import { StakePoolEpochRewards } from '../types';
 
 const MILLISECONDS_PER_DAY = 1000 * 60 * 60 * 24;