+++ packages/input-selection/src/index.ts
@@ -1,3 +1,4 @@
 export * from './RoundRobinRandomImprove';
+export * from './GreedySelection';
 export * from './types';
 export * from './InputSelectionError';