+++ packages/web-extension/src/observableWallet/util.ts
@@ -96,6 +96,7 @@ export const observableWalletProperties: RemoteApiProperties<ObservableWallet> =
   },
   currentEpoch$: RemoteApiPropertyType.HotObservable,
   delegation: {
+    distribution$: RemoteApiPropertyType.HotObservable,
     rewardAccounts$: RemoteApiPropertyType.HotObservable,
     rewardsHistory$: RemoteApiPropertyType.HotObservable
   },
@@ -104,6 +105,7 @@ export const observableWalletProperties: RemoteApiProperties<ObservableWallet> =
   finalizeTx: RemoteApiPropertyType.MethodReturningPromise,
   genesisParameters$: RemoteApiPropertyType.HotObservable,
   getName: RemoteApiPropertyType.MethodReturningPromise,
+  handles$: RemoteApiPropertyType.HotObservable,
   initializeTx: RemoteApiPropertyType.MethodReturningPromise,
   protocolParameters$: RemoteApiPropertyType.HotObservable,
   signData: RemoteApiPropertyType.MethodReturningPromise,